moving funcs to separate go file
